Broke some shit, fixed other shit. Fucking horse shit.
git-svn-id: https://jengerer.unfuddle.com/svn/jengerer_jim@45 e479ca42-947a-49ca-8679-5d91e2c49ab0

+++ Inventory.cpp
@@ -293,7 +293,7 @@ void Inventory::moveItem( Slot *source, Slot *destination )
 	destination->setItem( target );
 }
 
-bool Inventory::canMove ( uint8 index )
+bool Inventory::canMove ( uint16 index )
 {
 	int maxSlot = inventorySize();
 	if ((index >= 0) && (index < maxSlot)) {
